rename: sparseml_wrapper.py -> sparse_train_manager.py
